cms-admin: Extract EntityTypeChip and ScopeCell from GlobalActionLogGrid
Move the EntityTypeChip styled component to a GlobalActionLogGrid.sc.ts
sibling file and extract the scope column rendering (including the scope
label formatting) into a dedicated ScopeCell component.
